+++ models/Role.php
@@ -2,6 +2,6 @@
 
 class Role extends BaseModel
 {
-    protected string $table = 'VAI_TRO';
+    protected string $table = 'vai_tro';
     protected string $primaryKey = 'IdVaiTro';
 }